data/src/main/java/com/dyescape/bot/data/entity/data/UserEntity.java
package com.dyescape.bot.data.entity.data;

import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

import java.util.Objects;

@Entity
@Table(name = "discord_user")
public class UserEntity {

    @Id
    private String id;

    protected UserEntity() {

    }

    public UserEntity(String id) {
        this.id = id;
    }

    public String getId() {
        return this.id;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        UserEntity that = (UserEntity) o;
        return Objects.equals(id, that.id);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id);
    }
}
